Moved gift scheduler endpoints into one controller
ref https://linear.app/ghost/issue/BER-3851/establish-immediate-email-delivery-for-gift-subscriptions
The reminder and delivery wake-ups share the gifts API surface, so one controller avoids duplicate endpoint plumbing while preserving the existing route.
