+++ ghost/core/test/unit/api/endpoints/gifts.test.js
@@ -1,10 +1,10 @@
 const assert = require('node:assert/strict');
 const sinon = require('sinon');
 const domainEvents = require('@tryghost/domain-events');
-const giftRemindersController = require('../../../../core/server/api/endpoints/gift-reminders');
+const giftsController = require('../../../../core/server/api/endpoints/gifts');
 const StartGiftReminderFlushEvent = require('../../../../core/server/services/gifts/events/start-gift-reminder-flush-event');
 
-describe('Gift Reminders controller', function () {
+describe('Gifts controller', function () {
     afterEach(function () {
         sinon.restore();
     });
@@ -13,7 +13,7 @@ describe('Gift Reminders controller', function () {
         it('dispatches a StartGiftReminderFlushEvent', function () {
             const dispatchStub = sinon.stub(domainEvents, 'dispatch');
 
-            const result = giftRemindersController.flushReminders.query({});
+            const result = giftsController.flushReminders.query({});
 
             sinon.assert.calledOnceWithExactly(
                 dispatchStub,
